Change ft_putnbr_base to take long long instead of just long

+++ ft_putnbr_base.c
@@ -6,13 +6,13 @@
 /*   By: gsteyn <gsteyn@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/03 10:33:42 by gsteyn            #+#    #+#             */
-/*   Updated: 2018/08/03 10:51:19 by gsteyn           ###   ########.fr       */
+/*   Updated: 2018/08/06 13:16:17 by gsteyn           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void			ft_putnbr_base(long n, int base)
+void			ft_putnbr_base(long long n, int base)
 {
 	if (base < 2 || base > 16)
 		return ;
@@ -21,7 +21,7 @@ void			ft_putnbr_base(long n, int base)
 		ft_putchar('-');
 		n = -n;
 	}
-	if (n >= (long)base)
+	if (n >= (long long)base)
 	{
 		ft_putnbr_base(n / base, base);
 		ft_putchar(BASE[n % base]);
